physical-ai-backend/backend/pyproject.toml
[project]
name = "rag-content-ingestion"
version = "0.1.0"
description = "RAG Content Ingestion Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.12.0",
    "qdrant-client>=1.7.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"